api/requirements.txt
numpy>1.20.1
plotly==5.11.0
pyngrok==5.2.1
Pillow==9.3.0
streamlit==1.15.2
tensorflow-cpu
